refactor: extract header and nav drawer
